requirements.txt
# Core web framework
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
httpx==0.27.2
PyYAML==6.0.2

# UI Framework
gradio==4.44.1

# LLM Providers
openai>=1.0.0
google-generativeai==0.8.3
anthropic>=0.7.0

# Multi-agent orchestration
langgraph==0.2.39
langchain==0.3.7
langchain-openai==0.2.8

# Document processing
pypdf>=3.15.0,<5.0.0
python-docx>=1.1.0,<2.0.0
python-pptx>=0.6.21,<1.0.0
pandas>=2.2.0,<3.0.0
Pillow>=9.0.0,<11.0.0
pdf2image>=1.16.0,<2.0.0
PyMuPDF>=1.23.0,<2.0.0
openpyxl>=3.1.0,<4.0.0

# Utilities
requests>=2.28.0
aiofiles>=23.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0

# Audio generation for notifications
numpy>=1.25.0
